@@ -3,19 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/react": "11.1.0",
-    "@testing-library/user-event": "12.1.10",
+    "@testing-library/react": "16.2.0",
+    "@testing-library/user-event": "14.6.1",
     "@types/node": "20.8.4",
-    "@types/react": "17.0.0",
-    "@types/react-document-title": "2.0.4",
-    "@types/react-dom": "17.0.0",
-    "@types/styled-components": "5.1.2",
+    "@types/react": "19.0.10",
+    "@types/react-dom": "19.0.4",
+    "@types/styled-components": "5.1.34",
     "@vitejs/plugin-react-swc": "3.4.0",
-    "react": "17.0.2",
-    "react-document-title": "2.0.3",
-    "react-dom": "17.0.2",
+    "react": "19.0.0",
+    "react-document-title": "^2.0.3",
+    "react-dom": "19.0.0",
     "react-router-dom": "6.29.0",
-    "styled-components": "5.2.1",
+    "styled-components": "6.1.15",
     "typescript": "5.2.2",
     "vite": "4.4.11",
     "vite-plugin-svgr": "4.1.0",
