@@ -1,11 +1,9 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import AppRouter from "./routers/AppRouter";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <AppRouter />
-  </React.StrictMode>,
-  document.getElementById("root"),
-);
+const container = document.getElementById("root");
+const root = createRoot(container!);
+
+root.render(<AppRouter />);
