@@ -1,12 +1,13 @@
 import React from "react";
 import Page from "../../Page";
-import { useRef } from "react";
 
 const Frontpage = () => {
-  const ref = useRef<HTMLDivElement>(null);
-
   return (
-    <Page />
+    <Page
+      title={"Forside"}
+    >
+      <p>Hei, verden</p>
+    </Page>
   );
 };
 
